mkfs: add a check for conflicting values
Add a check that reports a conflict only when subopts are mixed with specific values.

For example, this should fail: -i attr=1 -m crc=1
But these should pass: -i attr=1 -m crc=0 or -i attr=2 -m crc=1
So we need a way how to raise conflict on these occassions. This does
not set the conflicting attributes in the options, but it prepares the
ground for it.

Signed-off-by: Jan Tulak <jtulak@redhat.com>