mkfs: add cross-section conflict checks
Checks are modified to work with cross-section conflicts (data, metada, log, ...).
So, it is now possible to detect a conflict between -m foo and -d bar
options.

This patch only extends checks to test for conflicts across all
options instead of only the current one, the opts struct already can
declare it.

Signed-off-by: Jan Tulak <jtulak@redhat.com>

--
changes:
* Whitespace cleaning
* Some whitespaces, conflict_struct() renamed to print_conflict_struct()
* Added a comment for the function ^^^
* And the ^^^ function was simplified to avoid copy&paste code

Signed-off-by: Jan Tulak <jtulak@redhat.com>