mkfs: save user input values into opts
Save the parsed values from users into the opts table. This will ensure
that the user values gets there and are validated, but we are not yet
using them for anything - the usage makes a lot of changes through the
code, so it is better if that is separated into smaller chunks.

Signed-off-by: Jan Tulak <jtulak@redhat.com>
Reviewed-by: Darrick J. Wong <darrick.wong@oracle.com>