mkfs: replace variables with opts table: -b,d,s options
Remove variables that can be replaced with a direct access to the opts
table, so we have it all in a single place, acessible from anywhere.

In future, we can remove some instances where we are passing values as
arguments to helper functions, when we have the values in the opts
struct and could pass only the struct.  But for now, limit the changes
to simple replacement without any change in the logic.

This is first of multiple similar patches that do the same, but for
other options.

Signed-off-by: Jan Tulak <jtulak@redhat.com>
Reviewed-by: Darrick J. Wong <darrick.wong@oracle.com>

---
CHANGE:
* indentation change mentioned by Darrick with reviewed-by